refactor(validators): move validation logic to middleware
Move validation logic from auth.routes.js, auth.validator.js to middleware

Refs: #30
