+++ src/api/components/auth/validators/auth.validator.js
@@ -1,24 +1,4 @@
 import * as Yup from 'yup';
-import APIError from '../../../utils/api-error.js';
-
-export const validate = (schema) => async (req, res, next) => {
-    try {
-        const data = await schema.validate(req.body, { abortEarly: false });
-        console.log('Validation successful:', data);
-        next();
-    } catch (error) {
-        console.error('Validation error:', error);
-        const apiError = new APIError({
-            message: 'Validation error',
-            status: 422,
-            isPublic: true,
-            stack: error.stack,
-            errorCode: 'validation_error',
-            errors: error.errors
-        });
-        next(apiError);
-    }
-};
 
 export const Schemas = {
     register: Yup.object().shape({
